refactors

+++ src/main/java/br/com/challenge_alura_one_t6/AluraForum/service/AnswerService.java
@@ -8,6 +8,7 @@ import br.com.challenge_alura_one_t6.AluraForum.exception.ObjectNotFoundExceptio
 import br.com.challenge_alura_one_t6.AluraForum.repositories.AnswerRepository;
 import br.com.challenge_alura_one_t6.AluraForum.repositories.TopicRepository;
 import br.com.challenge_alura_one_t6.AluraForum.repositories.UserRepository;
+import br.com.challenge_alura_one_t6.AluraForum.utils.IAuthenticationFacade;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
@@ -18,11 +19,13 @@ import java.util.Optional;
 @Transactional
 public class AnswerService {
     private final AnswerRepository answerRepository;
+    private final IAuthenticationFacade authenticationFacade;
     private final TopicRepository topicRepository;
     private final UserRepository userRepository;
 
-    public AnswerService(AnswerRepository answerRepository, TopicRepository topicRepository, UserRepository userRepository) {
+    public AnswerService(AnswerRepository answerRepository, IAuthenticationFacade authenticationFacade, TopicRepository topicRepository, UserRepository userRepository) {
         this.answerRepository = answerRepository;
+        this.authenticationFacade = authenticationFacade;
         this.topicRepository = topicRepository;
         this.userRepository = userRepository;
     }
@@ -37,11 +40,14 @@ public class AnswerService {
 
     }
     public Answer saveAnswer(AnswerDto answerDto){
-        Optional<User> optionalUser = userRepository.findById(answerDto.userId());
+       var name = authenticationFacade.getAuthentication().getName();
+       var user = userRepository.findByEmail(name);
+
         Optional<Topic> optionalTopic = topicRepository.findById(answerDto.topicId());
-        if(optionalUser.isPresent() && optionalTopic.isPresent()){
+
+        if(optionalTopic.isPresent()){
             Answer answer = new Answer();
-            answer.setUser(optionalUser.get());
+            answer.setUser((User) user);
             answer.setTopic(optionalTopic.get());
             answer.setMessage(answerDto.message());
             return answerRepository.save(answer);
@@ -50,14 +56,19 @@ public class AnswerService {
     }
 
     public Answer updateAnswer( Long answerId, AnswerDto answerDto) {
-        Optional<User> optionalUser = userRepository.findById(answerDto.userId());
+        var emailUser = authenticationFacade.getAuthentication().getName();
         Answer answer = answerRepository.findById(answerId).orElseThrow(()->new ObjectNotFoundException("answer",answerId));
+        var email = answer.getUser().getEmail();
+        if(emailUser != email)throw  new ObjectNotFoundException("Esta resposta não te pertence", answerId);
         answer.setMessage(answerDto.message());
         return answerRepository.save(answer);
     }
 
     public void deleteAnswer(Long answerId) {
-        answerRepository.findById(answerId).orElseThrow(()->new ObjectNotFoundException("answer",answerId));
+        var emailUser = authenticationFacade.getAuthentication().getName();
+        Answer answer = answerRepository.findById(answerId).orElseThrow(()->new ObjectNotFoundException("answer",answerId));
+        var email = answer.getUser().getEmail();
+        if(emailUser != email) throw  new ObjectNotFoundException("Voce não pode deletar este recurso",answerId);
         answerRepository.deleteById(answerId);
     }
 }
